Refactor RotationAbility: make a system
