--- Assets/Scripts/Field/Character/RotateAbility.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: ce4bd998362fca54f95e9a055571e2cb
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
