+++ Assets/Scripts/Systems/RotationSystem.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 5cfa9e75771449a185d45c7a03609e9c
+timeCreated: 1658407724
